Developed comprehensive code for course choices
Developed a series of methods to facilitate the displaying of current
course choices using the logged in CAO number. Also developed methods to
facilitate the updating of all course choices or just one course choice,
as suggested as a possible improvement in CA4 feedback. Created
complementary methods to enhance menu navigation too.