Parsed server response for single course & choices
Refactored the displaying of a single searched course and the current
student choices on the client to parse the necessary information from
the response received from the server in order to preserve client/server
encapsulation like previous commit. Outputed choices on a single line as
suggested in CA5 Deliverable 1 feedback.